requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
xgboost>=2.0.0
imbalanced-learn>=0.12.0
optuna>=3.5.0
shap>=0.44.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.19.0
streamlit>=1.30.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
joblib>=1.3.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.28.0
pytest>=8.0.0
